Plot a fitted vulnerability curve

Standard plots of fitted curves (objects returned by fitplc, fitplcs, fitcond or fitconds), with plenty of options for customization.

# S3 method for plcfit
plot(x, xlab = NULL, ylab = NULL, ylim = NULL,
  pch = 19, plotPx = TRUE, plotci = TRUE, plotdata = TRUE,
  plotfit = TRUE, add = FALSE, multiplier = NULL, px_ci = c("bootstrap",
  "parametric", "none"), px_ci_type = c("vertical", "horizontal"),
  px_ci_label = TRUE, plotrandom = FALSE, pointcol = "black",
  linecol = "black", linetype = 1, linelwd = 1, linecol2 = "blue",
  pxlinecol = "red", pxcex = 0.7, citype = c("polygon", "lines"),
  cicol = alpha("lightgrey", 0.8), what = c("relk", "PLC", "embol"),
  selines = NULL, xaxis = c("positive", "negative"), ...)

# S3 method for manyplcfit
plot(x, what = c("relk", "embol", "PLC"),
  onepanel = FALSE, linecol = NULL, pointcol = NULL, pch = 19,
  legend = TRUE, legendwhere = "topright", ...)

Arguments

x
A fitted curve returned by fitplc

xlab, ylab
Optionally, X and Y axis labels (if not provided, a default is used).

ylim
Optionally, Y-axis limits.

pch
Optionally, the plotting symbol (default = 19, filled circles)

plotPx
Logical (default TRUE), whether to plot a vertical line for the P50.

plotci
Logical (default TRUE), whether to plot the confidence interval (if computed with bootci=TRUE).

plotdata
Logical (default TRUE), whether to add the data to the plot.

plotfit
Logical (default TRUE), whether to add the fitted curve to the plot.

add
Logical (default FALSE), whether to add the plot to a current device. This is useful to overlay two plots or curves, for example.

multiplier
Multiply the scaled data (for plotting).

px_ci
Option for the confidence interval around Px, either 'parametric' (confidence interval computed with confint), 'bootstrap' (computed with non-parametric bootstrap) or 'none' (no plotting of the confidence interval) (formerly argument was called selines)

px_ci_type
Either 'vertical' (default), or 'horizontal', to plot confidence limits for Px.

px_ci_label
Logical (default TRUE), whether to write a label next to the CI for Px.

plotrandom
Logical. If TRUE (default is FALSE), plots the predictions for the random effects (only if random effects were included in the model fit).

pointcol
The color(s) of the data points.

linecol
The color(s) of the fitted curve (or color of the random effects curves if plotrandom=TRUE).

linetype
Line type for fitted curve (see options for lty in par).

linelwd
Width of the line (see options for lwd in par).

linecol2
The color of the fixed effects curve (if plotrandom=TRUE; otherwise ignored).

pxlinecol
The color of the lines indicating Px and its confidence interval

pxcex
Character size for the Px label above the Y-axis.

citype
Either 'polygon' (default), or 'lines', specifying formatting of the confidence interval in the plot.

cicol
The color of the confidence interval band (if plotted).

what
Either 'relk' or 'PLC' (or synonym 'embol'); it will plot either relative conductivity or percent loss conductivity (percent embolism).

selines
Obsolete; use px_ci

xaxis
Either 'positive' (default), so that water potential is plotted as positive values, or 'negative', plotting negative-valued water potentials.

…
Further parameters passed to plot, or points (when add=TRUE)

onepanel
For plotting of many curve fits, plot all curves in one panel (TRUE) or in separate panels (FALSE)

legend
(for fitconds and fitplcs only) Logical (default TRUE), whether to include a simple legend when plotting multiple fits

legendwhere
(for fitconds and fitplcs only) As in legend, specification of where to place legend (e.g. 'bottomleft'; coordinates not accepted)